environment support added

+++ adlsfolders/ps/Remove-Resources.ps1
@@ -4,16 +4,27 @@ param
     [String]$projectsParameterFile
 )
 
-function Remove-Resources {
-    $adlStoreName = $parameters.parameters.resources.value.adlStoreName
-
-    foreach ($obj in $parameters.parameters.resources.value.folders) {
+function Remove-Resource {
+    param(
+        [object]$resource
+    )
+    $adlStoreName = $resource.adlStoreName
+    $dataLakeStore = Get-AzureRmDataLakeStoreAccount | Where-Object {$_.Name -eq $adlStoreName}
+    if ($dataLakeStore -eq $null) {
+        Write-Verbose "Data Lake Store does not exist. Returning..."
+        return
+    }
+    foreach ($obj in $resource.folders) {
         if (-not ("/".Equals($obj.folderName))) {
             Try {
                 # Create the folder so long as it doesn't already exist.
                 if ((Get-AzureRmDataLakeStoreItem -AccountName $adlStoreName -Path $obj.folderName -ErrorAction Ignore)) {
+                    Write-Verbose "Remove folder $($obj.folderName)"
                     Remove-AzureRmDataLakeStoreItem -AccountName $adlStoreName -Path $obj.folderName -Recurse -Force
-                }            
+                }
+                else{
+                    Write-Verbose "Nothing to remove"
+                }
             }
             Catch {
                 Write-Error $_.Exception.Message
@@ -26,5 +37,4 @@ $parameterFileName = "$((Get-Item -Path $PSScriptRoot).Parent.Name).parameters.j
 $commonPSFolder = (Get-Item -Path "$PSScriptRoot\..\..\common\ps").FullName
 $null = & "$commonPSFolder\Invoke-RemoveProcess.ps1" `
     -projectsParameterFile $projectsParameterFile `
-    -parameterFileName $parameterFileName `
-    -procToRun {Remove-Resources}
+    -parameterFileName $parameterFileName
